+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Improvements**
 
+* Add `Command.consume_args()` for convenient argument handling.
 * Allow `@` in unquoted strings.
 
 # 0.3.0 (2024-05-25)